Кажется, справился
Страница погоды с viewPager-ом
Получение gps-координат
По координатам получение названия города
И, наконец, получение прогноза на 10 дней (из которых 7 оставляю)

+++ app/src/main/java/com/ultra/weatherforecast/Units/LatLng.java
@@ -9,11 +9,11 @@ package com.ultra.weatherforecast.Units;
 
 public class LatLng
 	{
-	public LatLng(float _lat,float _lng)
+	public LatLng(double _lat,double _lng)
 		{
 		lat=_lat;
 		lng=_lng;
 		}
 
-	public float lat,lng;
+	public double lat,lng;
 	}
